+++ spinn_front_end_common/interface/provenance/__init__.py
@@ -18,6 +18,7 @@ from .abstract_provides_local_provenance_data import (
 from .abstract_provides_provenance_data_from_machine import (
     AbstractProvidesProvenanceDataFromMachine)
 from .fec_timer import FecTimer
+from .global_provenance import GlobalProvenance
 from .log_store_db import LogStoreDB
 from .provenance_reader import ProvenanceReader
 from .provides_provenance_data_from_machine_impl import (
@@ -27,6 +28,7 @@ from .timer_category import TimerCategory
 from .timer_work import TimerWork
 
 __all__ = ["AbstractProvidesLocalProvenanceData", "FecTimer",
+           "GlobalProvenance",
            "AbstractProvidesProvenanceDataFromMachine", "LogStoreDB",
            "ProvenanceReader", "ProvenanceWriter",
            "ProvidesProvenanceDataFromMachineImpl",
